+++ src/core/methods/signTransactions/helpers/getGuardedTransactions.ts
@@ -1,9 +1,9 @@
 import { Transaction } from '@multiversx/sdk-core';
 import { getAreAllTransactionsSignedByGuardian } from './getAreAllTransactionsSignedByGuardian';
-import { ProviderFactory } from 'core/providers/ProviderFactory';
 import { getAccount } from 'core/methods/account/getAccount';
 import { walletAddressSelector } from 'store/selectors';
 import { getState } from 'store/store';
+import { createCrossWindowProvider } from 'core/providers/helpers/crossWindow/createCrossWindowProvider';
 
 export const getGuardedTransactions = async ({
   transactions
@@ -22,8 +22,7 @@ export const getGuardedTransactions = async ({
     return transactions;
   }
 
-  const factory = new ProviderFactory();
-  const provider = await factory.createCrossWindowProvider({
+  const provider = await createCrossWindowProvider({
     address,
     walletAddress
   });
